+++ module/dialogs/weapon-dialog.mjs
@@ -82,6 +82,7 @@ export function launchWeaponDialog({ weapon, attackType, actor }) {
 
   // Prepare data for the template
   const templateData = {
+    config : CONFIG.UNBOUNDFATE,
     weaponName,
     weaponDamage,
     attackType,